requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.2.0
Pillow==10.1.0
python-magic==0.4.27
gunicorn==21.2.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
marshmallow==3.19.0
